Refactor: Change stockAmount type from int to double
The `stockAmount` field in `ProductStockUpdateModel` and its usage in `ProductsAndServicesPage` have been updated from `int` to `double`.

This change allows for more precise stock management, accommodating fractional quantities.

+++ lib/Models/ProductsModels/ProductStockUpdateModel.dart
@@ -1,6 +1,6 @@
 class ProductStockUpdateModel {
   final int productId;
-  final int stockAmount;
+  final double stockAmount;
   final String description;
 
   ProductStockUpdateModel({
